move: helpers for TemplateDetails into that folder
isMinorRelease and TEMPLATE_SYSTEMS_UPDATE_LIMIT
were extracted from templateHelpers and
moved under TemplateDetails folder,
since there are the components there that import them.
